+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The following is a curated list of changes in the Enact project, newest changes 
 ### Changed
 
 - `ui/Scroller` scroll animation method for `scrollMode: native`
+## [5.3.1] - 2025-10-14
+
+### Fixed
+
+- `ui/Scroller` to calculate correctly targetX and targetY when scrolling to a position
+- `spotlight` navigation from an element to a different container
 
 ## [5.3.0] - 2025-09-24